Consolidate component information in llbooter, with FIXME: heap pointer is broken
- FIXME: The heap pointer is broken. We need to 1. setup the linker scipt to fix this, and (eventually) 2. update the heap pointer in the crt_comp_info page.
- Cleaned up a lot of the component bootup stuff by consolidating in llbooter.

+++ src/kernel/include/shared/cos_types.h
@@ -368,9 +368,9 @@ struct restartable_atomic_sequence {
 
 /* see explanation in spd.h */
 struct usr_inv_cap {
-	vaddr_t      invocation_fn;
-	unsigned int cap_no;
-	void        *data;
+	vaddr_t       invocation_fn;
+	unsigned long cap_no;
+	void         *data;
 };
 
 #define COMP_INFO_POLY_NUM 10
